% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.base.R
\docType{class}
\name{openmi.om.base-class}
\alias{openmi.om.base-class}
\alias{openmi.om.base}
\title{The base object class for meta-model components.}
\value{
reference class of type openmi.om.base.
}
\description{
The base object class for meta-model components.
}

