% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.equation.R
\docType{class}
\name{openmi.om.equation-class}
\alias{openmi.om.equation-class}
\alias{openmi.om.equation}
\title{The base class for executable equation based meta-model components.}
\value{
reference class of type openmi.om.equation
}
\description{
The base class for executable equation based meta-model components.
}

