% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.linkableComponent.R
\docType{class}
\name{openmi.om.linkableComponent-class}
\alias{openmi.om.linkableComponent-class}
\alias{openmi.om.linkableComponent}
\title{The base class for linkable meta-model components.}
\value{
reference class of type openmi.om.linkableComponent
}
\description{
The base class for linkable meta-model components.
}

