% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.runtimeController.R
\docType{class}
\name{openmi.om.runtimeController-class}
\alias{openmi.om.runtimeController-class}
\alias{openmi.om.runtimeController}
\title{The base class for meta-model simulation control.}
\value{
reference class of type openmi.om.runtimeController
}
\description{
The base class for meta-model simulation control.
}

