% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.timer.R
\docType{class}
\name{openmi.om.timer-class}
\alias{openmi.om.timer-class}
\alias{openmi.om.timer}
\title{The base time-keeping class for simulation control.}
\value{
reference class of type openmi.om.timer
}
\description{
The base time-keeping class for simulation control.
}

