% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.timer.R
\name{openmi.om.timer}
\alias{openmi.om.timer}
\title{The base time-keeping class for simulation control}
\value{
R6 class of type openmi.om.timer
}
\description{
Class providing the minimum attributes and methods for a model component
}
\details{
Has standard methods for iterating through timesteps and connecting with other components
}
\seealso{

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{starttime}}{beginning of simulation}

\item{\code{endtime}}{end of simulation}

\item{\code{thistime}}{current simulation time}

\item{\code{status}}{is this timer running, paused, finished}

\item{\code{mo}}{current simulation month}

\item{\code{da}}{current simulation day}

\item{\code{yr}}{current simulation year}

\item{\code{tz}}{current timezone}

\item{\code{dt}}{time step}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-update}{\code{openmi.om.timer$update()}}
\item \href{#method-init}{\code{openmi.om.timer$init()}}
\item \href{#method-clone}{\code{openmi.om.timer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
advance timer one timestep and update data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.timer$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-init"></a>}}
\if{latex}{\out{\hypertarget{method-init}{}}}
\subsection{Method \code{init()}}{
initialize for a model run
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.timer$init()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.timer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
