% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_mi_load.R
\name{openmi_om_load}
\alias{openmi_om_load}
\title{Parse a list object to instantiate a model component/sub-component}
\usage{
openmi_om_load(elem_list)
}
\arguments{
\item{elem_list}{list with model descriptors}
}
\value{
open.mi R6 class of matching type or generic R6 openmi.om.base
}
\description{
Parse a list object to instantiate a model component/sub-component
}
\examples{
NA
}
\seealso{
NA
}
